instrumentation-genai/opentelemetry-instrumentation-claude-agent-sdk/tests/requirements.oldest.txt
# Copyright The OpenTelemetry Authors
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# This variant of the requirements aims to test the system using
# the oldest supported version of external dependencies.

claude-agent-sdk>=0.1.0
pytest==7.4.4
pytest-asyncio==0.21.0
pyyaml>=6.0
wrapt==1.17.3
opentelemetry-exporter-otlp-proto-http~=1.30
opentelemetry-api==1.37
opentelemetry-sdk==1.37
opentelemetry-semantic-conventions==0.58b0
opentelemetry-instrumentation==0.58b0

opentelemetry-instrumentation-claude-agent-sdk
opentelemetry-util-genai>=0.2b0,<0.3b0

